requirements.txt
PyJWT
python-dotenv
requests
